refactor: extract streaming upload pipeline
