+++ modules/gaussian_splatting/core/gaussian_streaming.h
@@ -20,16 +20,17 @@
 class GaussianMemoryStream;
 class OS;
 class ResidencyBudgetController;
-class UploadSchedulerFacade;
 
 // Extracted subsystem headers (ISSUE-006 split)
 #include "streaming_quantization.h"
 #include "streaming_vram_regulator.h"
 #include "streaming_atlas.h"
+#include "streaming_upload_pipeline.h"
 
 // GPU memory streaming system with ring buffer for large datasets
 class GaussianStreamingSystem : public RefCounted {
     GDCLASS(GaussianStreamingSystem, RefCounted);
+    friend class StreamingUploadPipeline;
 
 public:
     static constexpr uint32_t CHUNK_SIZE = 65536;  // 64K splats per chunk
@@ -148,36 +149,6 @@ private:
         uint64_t frame_number = 0;
     };
 
-    struct PackJob {
-        uint32_t asset_id = PRIMARY_ASSET_ID;
-        uint32_t chunk_idx = UINT32_MAX;
-        uint32_t buffer_slot = UINT32_MAX;
-        uint32_t asset_generation = 0; // Captured at enqueue time for staleness detection
-        uint64_t enqueue_usec = 0;
-        uint32_t chunk_start = 0;
-        uint32_t chunk_count = 0;
-        bool uses_explicit_source_indices = false;
-        LocalVector<uint32_t> source_indices;
-        // Snapshot capture is deferred to worker threads to keep queueing lightweight.
-        Ref<GaussianData> data_ref;
-    };
-
-    struct PendingChunkUpload {
-        uint32_t asset_id = PRIMARY_ASSET_ID;
-        uint32_t chunk_idx = UINT32_MAX;
-        uint32_t buffer_slot = UINT32_MAX;
-        uint32_t asset_generation = 0; // For staleness detection
-        uint64_t enqueue_usec = 0;
-        Vector<PackedGaussian> packed_data;
-        SHCompressionMetrics metrics;
-        uint32_t bytes_uploaded = 0;
-    };
-
-    struct PackThreadContext {
-        GaussianStreamingSystem *system = nullptr;
-        uint32_t thread_index = 0;
-    };
-
     enum class EvictionResult {
         NoEviction,
         EvictedNonVisible,
@@ -185,12 +156,12 @@ private:
         SkippedAllVisible,
     };
 
-    // Debug counters for chunk culling
     struct ChunkCullingStats {
         uint32_t total_chunks = 0;
         uint32_t visible_chunks = 0;
         uint32_t frustum_culled_chunks = 0;
         uint32_t loaded_chunks = 0;
+
         void reset() {
             total_chunks = 0;
             visible_chunks = 0;
@@ -199,7 +170,6 @@ private:
         }
     };
 
-    // Camera velocity tracking for predictive prefetch
     struct CameraVelocityTracker {
         Vector3 last_position;
         Vector3 velocity;
@@ -211,7 +181,7 @@ private:
 
     struct VisibilityState {
         bool chunk_frustum_culling_enabled = true;
-        float chunk_frustum_padding = 1.5f;  // Conservative padding multiplier to prevent popping
+        float chunk_frustum_padding = 1.5f;
         float chunk_radius_multiplier = 1.0f;
         ChunkCullingStats culling_stats;
         LocalVector<uint32_t> visible_chunk_indices;
@@ -219,10 +189,10 @@ private:
         bool predictive_prefetch_enabled = true;
         float prefetch_lookahead_distance = 10.0f;
         LODBlendConfig lod_blend_config;
-        float current_lod_blend_factor = 1.0f;  // Global blend factor for the frame
-        int global_sh_band_level = 3;  // 0-3 for DC only to full SH
-        uint32_t lod_transitions_this_frame = 0; // Chunks whose LOD target/current level changed this update.
-        uint32_t prev_visible_count = 0;  // For change ratio calculation
+        float current_lod_blend_factor = 1.0f;
+        int global_sh_band_level = 3;
+        uint32_t lod_transitions_this_frame = 0;
+        uint32_t prev_visible_count = 0;
 
         void update_chunk_visibility(GaussianStreamingSystem &system, const Transform3D &camera_transform, const Projection &projection);
         bool is_chunk_in_frustum(const AABB &p_bounds, const Vector<Plane> &p_frustum_planes) const;
@@ -265,12 +235,13 @@ private:
             uint64_t last_used_frame = UINT64_MAX;
             float distance = 0.0f;
         };
-        uint64_t chunk_load_counter = 0;  // Strictly increasing counter for chunk loads
+
+        uint64_t chunk_load_counter = 0;
         uint32_t eviction_hysteresis_frames = 5;
         uint32_t max_evictions_per_frame = 4;
         uint32_t chunks_evicted_this_frame = 0;
-        uint32_t visible_chunks_evicted_this_frame = 0;  // Only visible chunks that were evicted
-        uint64_t last_stabilize_log_frame = 0;  // For throttling stabilize log messages
+        uint32_t visible_chunks_evicted_this_frame = 0;
+        uint64_t last_stabilize_log_frame = 0;
         uint64_t cached_eviction_frame = UINT64_MAX;
         uint64_t cached_non_primary_lru_frame = UINT64_MAX;
         uint32_t cached_non_primary_lru_cursor = 0;
@@ -281,240 +252,6 @@ private:
         EvictionResult evict_least_recently_used(GaussianStreamingSystem &system, bool p_allow_visible_eviction);
         bool evict_non_primary_lru(GaussianStreamingSystem &system);
         bool ensure_atlas_slot_available(GaussianStreamingSystem &system, uint32_t requesting_asset_id);
-    };
-
-    // Cache-line-aligned pack/upload telemetry counters.
-    // Groups 15 atomic counters that were previously scattered across
-    // UploadQueueState, eliminating false-sharing between pack worker
-    // threads and the main thread that samples analytics.
-    // Compiles to an empty struct in release builds (non-DEV_ENABLED).
-    struct alignas(64) PackTelemetry {
-#ifdef DEV_ENABLED
-        SafeFlag enabled;
-
-        // Pack timing
-        SafeNumeric<uint64_t> pack_time_usec_total;
-        SafeNumeric<uint64_t> pack_time_usec_max;
-        SafeNumeric<uint32_t> pack_jobs_completed;
-
-        // Upload volume
-        SafeNumeric<uint64_t> upload_bytes_total;
-        SafeNumeric<uint32_t> upload_chunks_completed;
-
-        // Pack-queue latency
-        SafeNumeric<uint64_t> pack_queue_latency_usec_total;
-        SafeNumeric<uint64_t> pack_queue_latency_usec_max;
-        SafeNumeric<uint32_t> pack_queue_latency_samples;
-
-        // Upload-queue latency
-        SafeNumeric<uint64_t> upload_queue_latency_usec_total;
-        SafeNumeric<uint64_t> upload_queue_latency_usec_max;
-        SafeNumeric<uint32_t> upload_queue_latency_samples;
-
-        // Mutex contention
-        SafeNumeric<uint64_t> pack_mutex_wait_usec_total;
-        SafeNumeric<uint64_t> pack_mutex_wait_usec_max;
-        SafeNumeric<uint32_t> pack_mutex_wait_samples;
-
-        // --- Convenience accessors (keep call-site API identical) ---
-
-        _ALWAYS_INLINE_ bool is_enabled() const { return enabled.is_set(); }
-        _ALWAYS_INLINE_ void set_enabled(bool p_on) { enabled.set_to(p_on); }
-
-        _ALWAYS_INLINE_ void add_pack_time(uint64_t p_usec) {
-            pack_time_usec_total.add(p_usec);
-            pack_jobs_completed.increment();
-            pack_time_usec_max.exchange_if_greater(p_usec);
-        }
-
-        _ALWAYS_INLINE_ void add_upload_bytes(uint64_t p_bytes) {
-            upload_bytes_total.add(p_bytes);
-        }
-
-        _ALWAYS_INLINE_ void add_upload_chunk() {
-            upload_chunks_completed.increment();
-        }
-
-        _ALWAYS_INLINE_ void add_pack_queue_latency(uint64_t p_usec) {
-            pack_queue_latency_usec_total.add(p_usec);
-            pack_queue_latency_samples.increment();
-            pack_queue_latency_usec_max.exchange_if_greater(p_usec);
-        }
-
-        _ALWAYS_INLINE_ void add_upload_queue_latency(uint64_t p_usec) {
-            upload_queue_latency_usec_total.add(p_usec);
-            upload_queue_latency_samples.increment();
-            upload_queue_latency_usec_max.exchange_if_greater(p_usec);
-        }
-
-        _ALWAYS_INLINE_ void add_mutex_wait(uint64_t p_usec) {
-            pack_mutex_wait_usec_total.add(p_usec);
-            pack_mutex_wait_samples.increment();
-            if (p_usec > 0) {
-                pack_mutex_wait_usec_max.exchange_if_greater(p_usec);
-            }
-        }
-
-        // Snapshot-and-reset used by the analytics sampler each frame.
-        struct Snapshot {
-            uint64_t pack_time_total = 0;
-            uint64_t pack_time_max = 0;
-            uint32_t pack_jobs = 0;
-            uint64_t upload_bytes = 0;
-            uint32_t upload_chunks = 0;
-            uint64_t pack_queue_lat_total = 0;
-            uint64_t pack_queue_lat_max = 0;
-            uint32_t pack_queue_lat_samples = 0;
-            uint64_t upload_queue_lat_total = 0;
-            uint64_t upload_queue_lat_max = 0;
-            uint32_t upload_queue_lat_samples = 0;
-            uint64_t mutex_wait_total = 0;
-            uint64_t mutex_wait_max = 0;
-            uint32_t mutex_wait_samples = 0;
-        };
-
-        // Atomically exchange every counter to zero and return the snapshot.
-        Snapshot exchange_and_reset();
-
-        // Non-destructive read (for debug logging that shouldn't drain counters).
-        Snapshot read_current() const;
-
-#else // !DEV_ENABLED -- every method compiles to nothing.
-        _ALWAYS_INLINE_ bool is_enabled() const { return false; }
-        _ALWAYS_INLINE_ void set_enabled(bool) {}
-        _ALWAYS_INLINE_ void add_pack_time(uint64_t) {}
-        _ALWAYS_INLINE_ void add_upload_bytes(uint64_t) {}
-        _ALWAYS_INLINE_ void add_upload_chunk() {}
-        _ALWAYS_INLINE_ void add_pack_queue_latency(uint64_t) {}
-        _ALWAYS_INLINE_ void add_upload_queue_latency(uint64_t) {}
-        _ALWAYS_INLINE_ void add_mutex_wait(uint64_t) {}
-
-        struct Snapshot {
-            uint64_t pack_time_total = 0;
-            uint64_t pack_time_max = 0;
-            uint32_t pack_jobs = 0;
-            uint64_t upload_bytes = 0;
-            uint32_t upload_chunks = 0;
-            uint64_t pack_queue_lat_total = 0;
-            uint64_t pack_queue_lat_max = 0;
-            uint32_t pack_queue_lat_samples = 0;
-            uint64_t upload_queue_lat_total = 0;
-            uint64_t upload_queue_lat_max = 0;
-            uint32_t upload_queue_lat_samples = 0;
-            uint64_t mutex_wait_total = 0;
-            uint64_t mutex_wait_max = 0;
-            uint32_t mutex_wait_samples = 0;
-        };
-
-        _ALWAYS_INLINE_ Snapshot exchange_and_reset() { return {}; }
-        _ALWAYS_INLINE_ Snapshot read_current() const { return {}; }
-#endif // DEV_ENABLED
-    };
-
-    struct UploadQueueState {
-        struct UploadBudgetState {
-            uint64_t upload_budget = 0;
-            uint64_t slice_limit = 0;
-            uint32_t completed_chunks = 0;
-            uint32_t chunk_limit = 0;
-        };
-        static constexpr uint32_t QUEUE_COMPACT_MIN_PREFIX = 128;
-
-        bool async_pack_enabled = true;
-        uint32_t pack_worker_threads = 2;
-        uint32_t max_pack_jobs_in_flight = 4;
-        uint32_t max_chunk_loads_per_frame = 16;
-        uint64_t max_upload_bytes_per_frame = 128 * 1024 * 1024;
-        uint64_t max_upload_bytes_per_slice = 16 * 1024 * 1024;
-        // Optional bandwidth cap. 0 disables token-bucket throttling.
-        uint64_t max_upload_bytes_per_second = 0;
-        // Token-bucket state for bandwidth-based upload throttling.
-        uint64_t upload_budget_tokens = 0;
-        uint64_t upload_budget_last_update_usec = 0;
-        uint32_t queued_chunk_loads_this_frame = 0;
-        std::atomic<uint32_t> pack_jobs_in_flight{0};
-        PackTelemetry telemetry;
-
-        // Last sampled pack/upload metrics (for performance monitors)
-        double last_pack_avg_ms = 0.0;
-        double last_pack_max_ms = 0.0;
-        uint32_t last_pack_jobs = 0;
-        double last_upload_mb = 0.0;
-        uint32_t last_upload_chunks = 0;
-        String cap_tier_preset = "custom";
-        bool cap_tier_active = false;
-        String cap_source_upload_mb_per_frame = "project_default";
-        String cap_source_upload_mb_per_slice = "project_default";
-        String cap_source_upload_mb_per_second = "project_default";
-        uint32_t effective_upload_cap_mb_per_frame = 128;
-        uint32_t effective_upload_cap_mb_per_slice = 16;
-        uint32_t effective_upload_cap_mb_per_second = 0;
-        bool upload_frame_cap_hit_this_frame = false;
-        bool upload_slice_cap_hit_this_frame = false;
-        bool upload_bandwidth_cap_hit_this_frame = false;
-        bool chunk_load_cap_hit_this_frame = false;
-        // Queue-pressure latch contract:
-        // - inactive => source=none, reason=none
-        // - active => source/reason are known contract tokens
-        // Enforced via StreamingQueuePressureController in queue/enqueue/process paths.
-        bool queue_pressure_active = false;
-        String queue_pressure_source = "none";
-        String queue_pressure_reason = "none";
-        double last_pack_queue_latency_avg_ms = 0.0;
-        double last_pack_queue_latency_max_ms = 0.0;
-        double last_upload_queue_latency_avg_ms = 0.0;
-        double last_upload_queue_latency_max_ms = 0.0;
-        double last_pack_mutex_wait_avg_ms = 0.0;
-        double last_pack_mutex_wait_max_ms = 0.0;
-
-        // Async pack thread state
-        Mutex pack_thread_lifecycle_mutex;
-        Mutex pack_mutex;
-        Semaphore pack_semaphore;
-        LocalVector<Thread *> pack_threads;
-        LocalVector<PackThreadContext> pack_thread_contexts;
-        std::atomic<bool> pack_thread_running{false};
-        std::atomic<bool> pack_thread_exit{false};
-        // Queue head cursors avoid O(n) front erases under sustained streaming pressure.
-        LocalVector<PackJob> pack_queue;
-        uint32_t pack_queue_read_idx = 0;
-        LocalVector<PendingChunkUpload *> upload_queue;
-        uint32_t upload_queue_read_idx = 0;
-        std::atomic<uint32_t> pack_queue_depth_cached{0};
-        std::atomic<uint32_t> upload_queue_depth_cached{0};
-
-        void start_pack_threads(GaussianStreamingSystem &system);
-        void stop_pack_threads(GaussianStreamingSystem &system);
-        void pack_thread_func(GaussianStreamingSystem &system, uint32_t p_thread_index);
-        bool queue_chunk_load(GaussianStreamingSystem &system, uint32_t asset_id, uint32_t chunk_idx);
-        void process_upload_queue(GaussianStreamingSystem &system);
-        void clear_pending_uploads(GaussianStreamingSystem &system);
-        void cancel_asset_jobs(GaussianStreamingSystem &system, uint32_t asset_id);
-        void cancel_chunk_jobs(GaussianStreamingSystem &system, uint32_t asset_id, uint32_t chunk_idx, uint32_t buffer_slot);
-        void load_streaming_tuning_config_from_project_settings(GaussianStreamingSystem &system);
-        bool has_pending_uploads();
-        bool pop_upload_job(PendingChunkUpload *&job);
-        UploadBudgetState prepare_upload_budget_state();
-        bool resolve_upload_chunk(GaussianStreamingSystem &system, PendingChunkUpload *job, StreamingChunk *&chunk);
-        bool upload_job_slices(GaussianStreamingSystem &system, RenderingDevice *submission_rd,
-                StreamingChunk &chunk, PendingChunkUpload *job, uint64_t total_bytes,
-                uint64_t &upload_budget, uint64_t slice_limit, bool &submitted);
-        void finalize_upload_job(GaussianStreamingSystem &system, PendingChunkUpload *job,
-                StreamingChunk &chunk, UploadBudgetState &budget);
-        void requeue_upload_job(PendingChunkUpload *job);
-        uint32_t get_pack_queue_depth_cached() const;
-        uint32_t get_upload_queue_depth_cached() const;
-        void get_pending_queue_depths_cached(uint32_t &r_pack_queue_depth, uint32_t &r_upload_queue_depth) const;
-        uint32_t get_pack_queue_depth_unsafe() const;
-        uint32_t get_upload_queue_depth_unsafe() const;
-        void compact_queues_locked();
-        void sync_cached_queue_depths_locked();
-        void record_pack_mutex_wait(uint64_t wait_start_usec);
-        void record_upload_queue_latency(uint64_t enqueue_usec);
-
-    private:
-        bool start_pack_threads_locked(GaussianStreamingSystem &system);
-        void stop_pack_threads_locked(GaussianStreamingSystem &system);
     };
 
     struct BudgetState {
@@ -540,7 +277,7 @@ private:
     VisibilityState visibility;
     ZeroVisibleRecoveryState zero_visible_recovery;
     EvictionState eviction;
-    UploadQueueState uploads;
+    StreamingUploadPipeline upload_pipeline;
     BudgetState budget;
 
     struct AssetRegistryState {
@@ -878,7 +615,7 @@ public:
     uint32_t get_effective_splat_count() const;  // Total splats after LOD reduction
 
     // Internal upload scheduler access (not user-facing API).
-    UploadQueueState &_internal_get_upload_state() { return uploads; }
+    StreamingUploadPipeline &_internal_get_upload_pipeline() { return upload_pipeline; }
 
 private:
     bool _create_chunks();
@@ -924,6 +661,7 @@ private:
             GaussianSplatManager::ScopedSubmissionLock &submission_lock) const;
     bool _pack_chunk_data(uint32_t asset_id, uint32_t chunk_idx, const AtlasAssetState &asset, StreamingChunk &chunk,
             Vector<PackedGaussian> &chunk_data, SHCompressionMetrics &metrics);
+    void _complete_chunk_load_common(uint32_t asset_id, uint32_t chunk_idx, StreamingChunk &chunk);
     void _log_chunk_load_metrics(uint32_t chunk_idx, const SHCompressionMetrics &metrics);
     bool _upload_chunk_to_gpu(RenderingDevice *submission_rd, uint32_t buffer_offset,
             const Vector<PackedGaussian> &chunk_data, uint32_t asset_id, uint32_t chunk_idx,
@@ -942,8 +680,6 @@ private:
     void _load_streaming_tuning_config_from_project_settings();
     void _start_pack_threads();
     void _stop_pack_threads();
-    void _pack_thread_func(uint32_t p_thread_index);
-    static void _pack_thread_entry(void *p_userdata);
     bool _queue_chunk_load(uint32_t chunk_idx);
     bool _queue_chunk_load(uint32_t asset_id, uint32_t chunk_idx);
     bool _enqueue_chunk_load_request(uint32_t asset_id, uint32_t chunk_idx,
